+++ src/main/java/fuegoquasar/starwars/models/Position.java
@@ -4,26 +4,26 @@ package fuegoquasar.starwars.models;
 public class Position {
 
     private long id;
-    private double x;
-    private double y;
+    private float x;
+    private float y;
     
-    public Position(double x, double y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
     public long getId() {
         return id;
     }
-    public double getX() {
+    public float getX() {
         return x;
     }
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
-    public double getY() {
+    public float getY() {
         return y;
     }
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
